Remove requestrc, implement behaviors
Te request record collector was redundant with the better generic
api rc. It was also the only one that didn't use the base rc, which
made it the only one that couldn't be nade to handle behaviors
automatically.

Behaviors control _how_ requests are made. For now the only implemented
setting is `parallelLimit`, which sets a limit on the number of parallel
requests ec will make for one non-tree dependency.